safari-600.7-branch/LayoutTests/accessibility/range-alter-by-step-expected.txt


This test makes sure that if a range type has a step value, that increment and decrement work.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.intValue is 25
PASS obj.intValue is 50
PASS obj.intValue is 75
PASS obj.intValue is 100
PASS obj.intValue is 100
PASS obj.intValue is 75
PASS obj.intValue is 50
PASS obj.intValue is 25
PASS obj.intValue is 0
PASS obj.intValue is 0
PASS successfullyParsed is true

TEST COMPLETE
